Migrate ColorPickerModal to top level
The color picker modal was nested a few levels into the HTML - unfortunately, when changing the table provider for the EditableTable, a parent element started using the `transform` property for CSS.  This interferes with the fixed position that the modal was using.  As such, it needed to be migrated farther up the tree.  To facilitate this, there is now a service for interacting with it.